// src/types/workspace.ts
export type PreviewMode = "real" | "demo" | "fake";

export type ResponsiveDevice = "desktop" | "tablet" | "mobile";